fix: change logic to compute dependability
now it is based on actual difference of milliseconds, not only on
difference of days

+++ src/main/java/it/polito/ezgas/scheduling/ScheduledTasks.java
@@ -85,7 +85,7 @@ public class ScheduledTasks {
 		//    obsolescence = 0 if (today - P.time_tag) > 7 days 
 		//    otherwise obsolescence = 1 - (today - P.time_tag)/7
 		long msecDifference = this.now.getTime() - reportTimestamp.getTime();
-		int daysDifference = (int)(msecDifference / (24 * 60 * 60 * 1000));
+		double daysDifference = msecDifference / (24 * 60 * 60 * 1000.0);
 		double obsolescence = daysDifference > 7 ? 0 : 1 - daysDifference / 7.0;
 
 		User reportUser;
